// src/navigation/types.ts
export type VehiclesStackParamList = {
  VehicleList:       undefined;
  VehicleDetail:     { vehicleNo: string; vehicle?: VehicleDetailPayload };
  VehicleGroupList:  undefined;
  VehicleGroupDetail:{ groupId: number };
};